# api/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.CreateRestaurantRequest:
    properties:
      address:
        type: string
      description:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  genproto.RestaurantResponse:
    properties:
      address:
        type: string
      description:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  genproto.RestaurantsResponse:
    properties:
      restaurants:
        items:
          $ref: '#/definitions/genproto.RestaurantResponse'
        type: array
    type: object
...
  genproto.UpdateRestaurantRequest:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: APi service
  version: "1.0"
paths:
...
  /api/restaurant/create:
    post:
      consumes:
      - application/json
      description: Create a new restaurant
      parameters:
      - description: Create Restaurant
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genproto.CreateRestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.RestaurantResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Restaurant
      tags:
      - Restaurant
  /api/restaurant/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Restaurant
      tags:
      - Restaurant
  /api/restaurant/get-all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all restaurants
      parameters:
      - in: query
        name: address
        type: string
      - in: query
        name: description
        type: string
      - in: query
        name: name
        type: string
      - in: query
        name: phone_number
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.RestaurantsResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get All Restaurants
      tags:
      - Restaurant
  /api/restaurant/get-by-id/{id}:
    get:
      consumes:
      - application/json
      description: Get a restaurant by its ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.RestaurantResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Restaurant by ID
      tags:
      - Restaurant
  /api/restaurant/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing restaurant
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Restaurant
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genproto.UpdateRestaurantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.RestaurantResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Restaurant
      tags:
      - Restaurant
swagger: "2.0"